UI_test/requirements.txt
allure-pytest==2.10.0
allure-python-commons==2.10.0
async-generator==1.10
atomicwrites==1.4.1
attrs==22.1.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
execnet==1.9.0
h11==0.13.0
idna==3.3
iniconfig==1.1.1
outcome==1.2.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.2
pytest-forked==1.4.0
pytest-xdist==2.5.0
requests==2.28.1
selenium==4.4.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.12
wsproto==1.2.0
